Move action setupTooltip to a separate file
