+++ packages/plugins/Gfi/src/store/actions/index.ts
@@ -1,18 +1,17 @@
 import debounce from 'lodash.debounce'
 import { Style, Fill, Stroke } from 'ol/style'
-import Overlay from 'ol/Overlay'
 import { GeoJSON } from 'ol/format'
 import { Feature } from 'ol'
 import { Feature as GeoJsonFeature, GeoJsonProperties } from 'geojson'
 import { PolarActionTree } from '@polar/lib-custom-types'
 import getCluster from '@polar/lib-get-cluster'
-import { getTooltip, Tooltip } from '@polar/lib-tooltip'
 import { DragBox } from 'ol/interaction'
 import { platformModifierKeyOnly } from 'ol/events/condition'
 import { getFeatureDisplayLayer, clear } from '../../utils/displayFeatureLayer'
 import { GfiGetters, GfiState } from '../../types'
 import { getOriginalFeature } from '../../utils/getOriginalFeature'
 import { debouncedGfiRequest } from './debouncedGfiRequest'
+import { setupTooltip } from './setup'
 
 // OK for module action set creation
 // eslint-disable-next-line max-lines-per-function
@@ -143,55 +142,7 @@ export const makeActions = () => {
         )
       }
     },
-    setupTooltip({ getters: { gfiConfiguration }, rootGetters: { map } }) {
-      const tooltipLayerIds = Object.keys(gfiConfiguration.layers).filter(
-        (key) => gfiConfiguration.layers[key].showTooltip
-      )
-      if (!tooltipLayerIds.length) {
-        return
-      }
-
-      let element: Tooltip['element'], unregister: Tooltip['unregister']
-      const overlay = new Overlay({
-        positioning: 'bottom-center',
-        offset: [0, -5],
-      })
-      map.addOverlay(overlay)
-      map.on('pointermove', ({ pixel, dragging, originalEvent }) => {
-        if (dragging || ['touch', 'pen'].includes(originalEvent.pointerType)) {
-          return
-        }
-        let hasFeatureAtPixel = false
-        // stops on return `true`, thus only using the uppermost feature
-        map.forEachFeatureAtPixel(
-          pixel,
-          (feature, layer) => {
-            if (!(feature instanceof Feature)) {
-              return false
-            }
-            hasFeatureAtPixel = true
-            overlay.setPosition(map.getCoordinateFromPixel(pixel))
-            if (unregister) {
-              unregister()
-            }
-            ;({ element, unregister } = getTooltip({
-              localeKeys:
-                // @ts-expect-error | it exists by virtue of layerFilter below
-                gfiConfiguration.layers[layer.get('id')].showTooltip(
-                  feature,
-                  map
-                ),
-            }))
-            overlay.setElement(element)
-            return true
-          },
-          { layerFilter: (layer) => tooltipLayerIds.includes(layer.get('id')) }
-        )
-        if (!hasFeatureAtPixel) {
-          overlay.setPosition(undefined)
-        }
-      })
-    },
+    setupTooltip,
     setupFeatureVisibilityUpdates({ commit, state, getters, rootGetters }) {
       // debounce to prevent update spam
       debouncedVisibilityChangeIndicator = debounce(
